Add Alignment::Stretch and improve flex example
Alignment::Stretch forces children to occupy the maximum space on
the minor axis.

The more interesting part of this patch is that examples/flex
(previously examples/align) now allows the user to select the
settings for the demo widget, and updates in real-time; this is
useful for me for looking for evaluating layout issues.